remove nft api from internal
