+++ src/api/nft-namespace.ts
@@ -1,4 +1,4 @@
-import { BigNumberish } from '@ethersproject/bignumber';
+import { BigNumber, BigNumberish } from '@ethersproject/bignumber';
 import {
   GetBaseNftsForContractOptions,
   GetBaseNftsForOwnerOptions,
@@ -14,26 +14,33 @@ import {
   OwnedBaseNftsResponse,
   OwnedNft,
   OwnedNftsResponse,
-  RefreshContractResult
+  RefreshContractResult,
+  RefreshState
 } from '../types/types';
 import { BaseNft, Nft, NftContract } from './nft';
-import {
-  checkNftOwnership,
-  getNftContractMetadata,
-  getFloorPrice,
-  getNftMetadata,
-  getNftsForContract,
-  getNftsForNftContractIterator,
-  getNftsForOwner,
-  getNftsForOwnerIterator,
-  getOwnersForNft,
-  getOwnersForContract,
-  getSpamContracts,
-  isSpamContract,
-  refreshNftContract,
-  refreshNftMetadata
-} from '../internal/nft-api';
 import { AlchemyConfig } from './alchemy-config';
+import { paginateEndpoint, requestHttpWithBackoff } from '../internal/dispatch';
+import {
+  RawBaseNft,
+  RawGetBaseNftsForNftContractResponse,
+  RawGetBaseNftsResponse,
+  RawGetNftsForNftContractResponse,
+  RawGetNftsResponse,
+  RawGetOwnersForNftContractResponse,
+  RawNft,
+  RawNftContract,
+  RawNftContractBaseNft,
+  RawNftContractNft,
+  RawOwnedBaseNft,
+  RawOwnedNft,
+  RawReingestContractResponse
+} from '../internal/raw-interfaces';
+import { AlchemyApiType } from '../util/const';
+import {
+  getBaseNftFromRaw,
+  getNftContractFromRaw,
+  getNftFromRaw
+} from '../util/util';
 
 export class NftNamespace {
   constructor(private readonly config: AlchemyConfig) {}
@@ -46,12 +53,22 @@ export class NftNamespace {
    * @param tokenType - Optionally specify the type of token to speed up the query.
    * @public
    */
-  getNftMetadata(
+  async getNftMetadata(
     contractAddress: string,
     tokenId: BigNumberish,
     tokenType?: NftTokenType
   ): Promise<Nft> {
-    return getNftMetadata(this.config, contractAddress, tokenId, tokenType);
+    const response = await requestHttpWithBackoff<GetNftMetadataParams, RawNft>(
+      this.config,
+      AlchemyApiType.NFT,
+      'getNFTMetadata',
+      {
+        contractAddress,
+        tokenId: BigNumber.from(tokenId!).toString(),
+        tokenType: tokenType !== NftTokenType.UNKNOWN ? tokenType : undefined
+      }
+    );
+    return getNftFromRaw(response, contractAddress);
   }
 
   /**
@@ -60,8 +77,15 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT.
    * @public
    */
-  getNftContractMetadata(contractAddress: string): Promise<NftContract> {
-    return getNftContractMetadata(this.config, contractAddress);
+  async getNftContractMetadata(contractAddress: string): Promise<NftContract> {
+    const response = await requestHttpWithBackoff<
+      GetNftContractMetadataParams,
+      RawNftContract
+    >(this.config, AlchemyApiType.NFT, 'getContractMetadata', {
+      contractAddress
+    });
+
+    return getNftContractFromRaw(response);
   }
 
   /**
@@ -92,11 +116,34 @@ export class NftNamespace {
     owner: string,
     options?: GetBaseNftsForOwnerOptions
   ): AsyncIterable<OwnedBaseNft>;
-  getNftsForOwnerIterator(
+  async *getNftsForOwnerIterator(
     owner: string,
     options?: GetNftsForOwnerOptions | GetBaseNftsForOwnerOptions
   ): AsyncIterable<OwnedBaseNft | OwnedNft> {
-    return getNftsForOwnerIterator(this.config, owner, options);
+    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
+    for await (const response of paginateEndpoint(
+      this.config,
+      AlchemyApiType.NFT,
+      'getNFTs',
+      'pageKey',
+      'pageKey',
+      {
+        contractAddresses: options?.contractAddresses,
+        pageKey: options?.pageKey,
+        filters: options?.excludeFilters,
+        owner,
+        withMetadata
+      }
+    )) {
+      for (const ownedNft of response.ownedNfts as
+        | RawOwnedNft[]
+        | RawOwnedBaseNft[]) {
+        yield {
+          ...nftFromGetNftResponse(ownedNft),
+          balance: parseInt(ownedNft.balance)
+        };
+      }
+    }
   }
 
   /**
@@ -127,11 +174,29 @@ export class NftNamespace {
     owner: string,
     options?: GetBaseNftsForOwnerOptions
   ): Promise<OwnedBaseNftsResponse>;
-  getNftsForOwner(
+  async getNftsForOwner(
     owner: string,
     options?: GetNftsForOwnerOptions | GetBaseNftsForOwnerOptions
   ): Promise<OwnedNftsResponse | OwnedBaseNftsResponse> {
-    return getNftsForOwner(this.config, owner, options);
+    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
+    const response = await requestHttpWithBackoff<
+      GetNftsAlchemyParams,
+      RawGetBaseNftsResponse | RawGetNftsResponse
+    >(this.config, AlchemyApiType.NFT, 'getNFTs', {
+      contractAddresses: options?.contractAddresses,
+      pageKey: options?.pageKey,
+      filters: options?.excludeFilters,
+      owner,
+      withMetadata
+    });
+    return {
+      ownedNfts: response.ownedNfts.map(res => ({
+        ...nftFromGetNftResponse(res),
+        balance: parseInt(res.balance)
+      })),
+      pageKey: response.pageKey,
+      totalCount: response.totalCount
+    };
   }
 
   /**
@@ -163,11 +228,26 @@ export class NftNamespace {
     contractAddress: string,
     options?: GetBaseNftsForContractOptions
   ): Promise<NftContractBaseNftsResponse>;
-  getNftsForContract(
+  async getNftsForContract(
     contractAddress: string,
     options?: GetBaseNftsForContractOptions | GetNftsForContractOptions
   ): Promise<NftContractNftsResponse | NftContractBaseNftsResponse> {
-    return getNftsForContract(this.config, contractAddress, options);
+    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
+    const response = await requestHttpWithBackoff<
+      GetNftsForNftContractAlchemyParams,
+      RawGetBaseNftsForNftContractResponse | RawGetNftsForNftContractResponse
+    >(this.config, AlchemyApiType.NFT, 'getNFTsForCollection', {
+      contractAddress,
+      startToken: options?.pageKey,
+      withMetadata
+    });
+
+    return {
+      nfts: response.nfts.map(res =>
+        nftFromGetNftNftContractResponse(res, contractAddress)
+      ),
+      pageKey: response.nextToken
+    };
   }
 
   /**
@@ -201,11 +281,29 @@ export class NftNamespace {
     contractAddress: string,
     options?: GetBaseNftsForContractOptions
   ): AsyncIterable<BaseNft>;
-  getNftsForContractIterator(
+  async *getNftsForContractIterator(
     contractAddress: string,
     options?: GetBaseNftsForContractOptions | GetNftsForContractOptions
   ): AsyncIterable<BaseNft | Nft> {
-    return getNftsForNftContractIterator(this.config, contractAddress, options);
+    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
+    for await (const response of paginateEndpoint(
+      this.config,
+      AlchemyApiType.NFT,
+      'getNFTsForCollection',
+      'startToken',
+      'nextToken',
+      {
+        contractAddress,
+        startToken: options?.pageKey,
+        withMetadata
+      }
+    )) {
+      for (const nft of response.nfts as
+        | RawNftContractBaseNft[]
+        | RawNftContractNft[]) {
+        yield nftFromGetNftNftContractResponse(nft, contractAddress);
+      }
+    }
   }
 
   /**
@@ -214,10 +312,19 @@ export class NftNamespace {
    * @param contractAddress - The NFT contract to get the owners for.
    * @beta
    */
-  getOwnersForContract(
+  async getOwnersForContract(
     contractAddress: string
   ): Promise<GetOwnersForContractResponse> {
-    return getOwnersForContract(this.config, contractAddress);
+    const response = await requestHttpWithBackoff<
+      GetOwnersForNftContractAlchemyParams,
+      RawGetOwnersForNftContractResponse
+    >(this.config, AlchemyApiType.NFT, 'getOwnersForCollection', {
+      contractAddress
+    });
+
+    return {
+      owners: response.ownerAddresses
+    };
   }
 
   /**
@@ -227,11 +334,19 @@ export class NftNamespace {
    * @param tokenId - Token id of the NFT.
    * @beta
    */
-  getOwnersForNft(
+  async getOwnersForNft(
     contractAddress: string,
     tokenId: BigNumberish
   ): Promise<GetOwnersForNftResponse> {
-    return getOwnersForNft(this.config, contractAddress, tokenId);
+    return requestHttpWithBackoff(
+      this.config,
+      AlchemyApiType.NFT,
+      'getOwnersForToken',
+      {
+        contractAddress,
+        tokenId: BigNumber.from(tokenId!).toString()
+      }
+    );
   }
 
   /**
@@ -241,11 +356,18 @@ export class NftNamespace {
    * @param contractAddresses - An array of NFT contract addresses to check ownership for.
    * @beta
    */
-  checkNftOwnership(
+  async checkNftOwnership(
     owner: string,
     contractAddresses: string[]
   ): Promise<boolean> {
-    return checkNftOwnership(this.config, owner, contractAddresses);
+    if (contractAddresses.length === 0) {
+      throw new Error('Must provide at least one contract address');
+    }
+    const response = await this.getNftsForOwner(owner, {
+      contractAddresses,
+      omitMetadata: true
+    });
+    return response.ownedNfts.length > 0;
   }
 
   /**
@@ -257,7 +379,14 @@ export class NftNamespace {
    * @beta
    */
   isSpamContract(contractAddress: string): Promise<boolean> {
-    return isSpamContract(this.config, contractAddress);
+    return requestHttpWithBackoff<IsSpamContractParams, boolean>(
+      this.config,
+      AlchemyApiType.NFT,
+      'isSpamContract',
+      {
+        contractAddress
+      }
+    );
   }
 
   /**
@@ -268,7 +397,12 @@ export class NftNamespace {
    * @beta
    */
   getSpamContracts(): Promise<string[]> {
-    return getSpamContracts(this.config);
+    return requestHttpWithBackoff<undefined, string[]>(
+      this.config,
+      AlchemyApiType.NFT,
+      'getSpamContracts',
+      undefined
+    );
   }
 
   /**
@@ -278,7 +412,14 @@ export class NftNamespace {
    * @beta
    */
   getFloorPrice(contractAddress: string): Promise<GetFloorPriceResponse> {
-    return getFloorPrice(this.config, contractAddress);
+    return requestHttpWithBackoff<GetFloorPriceParams, GetFloorPriceResponse>(
+      this.config,
+      AlchemyApiType.NFT,
+      'getFloorPrice',
+      {
+        contractAddress
+      }
+    );
   }
 
   /**
@@ -296,11 +437,14 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT.
    * @param tokenId - The token id of the NFT.
    */
-  refreshNftMetadata(
+  async refreshNftMetadata(
     contractAddress: string,
     tokenId: BigNumberish
   ): Promise<boolean> {
-    return refreshNftMetadata(this.config, contractAddress, tokenId);
+    const tokenIdString = BigNumber.from(tokenId!).toString();
+    const first = await this.getNftMetadata(contractAddress, tokenIdString);
+    const second = await refresh(this.config, contractAddress, tokenIdString);
+    return first.timeLastUpdated !== second.timeLastUpdated;
   }
 
   /**
@@ -314,7 +458,189 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT collection.
    * @beta
    */
-  refreshContract(contractAddress: string): Promise<RefreshContractResult> {
-    return refreshNftContract(this.config, contractAddress);
+  async refreshContract(
+    contractAddress: string
+  ): Promise<RefreshContractResult> {
+    const response = await requestHttpWithBackoff<
+      ReingestContractParams,
+      RawReingestContractResponse
+    >(this.config, AlchemyApiType.NFT, 'reingestContract', {
+      contractAddress
+    });
+
+    return {
+      contractAddress: response.contractAddress,
+      refreshState: parseReingestionState(response.reingestionState),
+      progress: response.progress
+    };
   }
+}
+
+async function refresh(
+  config: AlchemyConfig,
+  contractAddress: string,
+  tokenId: BigNumberish
+): Promise<Nft> {
+  const response = await requestHttpWithBackoff<GetNftMetadataParams, RawNft>(
+    config,
+    AlchemyApiType.NFT,
+    'getNFTMetadata',
+    {
+      contractAddress,
+      tokenId: BigNumber.from(tokenId!).toString(),
+      refreshCache: true
+    }
+  );
+  return getNftFromRaw(response, contractAddress);
+}
+
+/**
+ * Helper method to convert a NFT response received from Alchemy backend to an
+ * SDK NFT type.
+ *
+ * @internal
+ */
+function nftFromGetNftResponse(
+  ownedNft: RawOwnedBaseNft | RawOwnedNft
+): Nft | BaseNft {
+  if (isNftWithMetadata(ownedNft)) {
+    return getNftFromRaw(ownedNft, ownedNft.contract.address);
+  } else {
+    return getBaseNftFromRaw(ownedNft, ownedNft.contract.address);
+  }
+}
+
+/**
+ * Helper method to convert a NFT response received from Alchemy backend to an
+ * SDK NFT type.
+ *
+ * @internal
+ */
+function nftFromGetNftNftContractResponse(
+  ownedNft: RawNftContractBaseNft | RawNftContractNft,
+  contractAddress: string
+): Nft | BaseNft {
+  if (isNftWithMetadata(ownedNft)) {
+    return getNftFromRaw(ownedNft, contractAddress);
+  } else {
+    return getBaseNftFromRaw(ownedNft, contractAddress);
+  }
+}
+
+/** @internal */
+// TODO: more comprehensive type check
+function isNftWithMetadata(response: RawBaseNft | RawNft): response is RawNft {
+  return (response as RawNft).title !== undefined;
+}
+
+/**
+ * Flips the `omitMetadata` SDK parameter type to the `withMetadata` parameter
+ * required by the Alchemy API. If `omitMetadata` is undefined, the SDK defaults
+ * to including metadata.
+ *
+ * @internal
+ */
+function omitMetadataToWithMetadata(
+  omitMetadata: boolean | undefined
+): boolean {
+  return omitMetadata === undefined ? true : !omitMetadata;
+}
+
+function parseReingestionState(reingestionState: string): RefreshState {
+  switch (reingestionState) {
+    case 'does_not_exist':
+      return RefreshState.DOES_NOT_EXIST;
+    case 'already_queued':
+      return RefreshState.ALREADY_QUEUED;
+    case 'in_progress':
+      return RefreshState.IN_PROGRESS;
+    case 'finished':
+      return RefreshState.FINISHED;
+    case 'queued':
+      return RefreshState.QUEUED;
+    case 'queue_failed':
+      return RefreshState.QUEUE_FAILED;
+    default:
+      throw new Error('Unknown reingestion state: ' + reingestionState);
+  }
+}
+
+/**
+ * Interface for the `getNftsForNftContract` endpoint. The main difference is
+ * that the endpoint has a `startToken` parameter, but the SDK standardizes all
+ * pagination parameters to `pageKey`.
+ *
+ * @internal
+ */
+interface GetNftsForNftContractAlchemyParams {
+  contractAddress: string;
+  startToken?: string;
+  withMetadata: boolean;
+}
+
+/**
+ * Interface for the `getNfts` endpoint. The main difference is that the
+ * endpoint has a `withMetadata` parameter, but the SDK exposes the parameter as
+ * `omitMetadata`.
+ *
+ * @internal
+ */
+interface GetNftsAlchemyParams {
+  owner: string;
+  pageKey?: string;
+  contractAddresses?: string[];
+  filters?: string[];
+  withMetadata: boolean;
+}
+
+/**
+ * Interface for the `getNftMetadata` endpoint.
+ *
+ * @internal
+ */
+interface GetNftMetadataParams {
+  contractAddress: string;
+  tokenId: string;
+  tokenType?: NftTokenType;
+  refreshCache?: boolean;
+}
+
+/**
+ * Interface for the `isSpamContract` endpoint.
+ *
+ * @internal
+ */
+interface IsSpamContractParams {
+  contractAddress: string;
+}
+
+/**
+ * Interface for the `getNftContractMetadata` endpoint.
+ *
+ * @internal
+ */
+interface GetNftContractMetadataParams {
+  contractAddress: string;
+}
+
+/**
+ * Interface for the `getOwnersForNftContract` endpoint.
+ *
+ * @internal
+ */
+interface GetOwnersForNftContractAlchemyParams {
+  contractAddress: string;
+}
+
+/**
+ * Interface for the `getFloorPrice` endpoint.
+ *
+ * @internal
+ */
+interface GetFloorPriceParams {
+  contractAddress: string;
+}
+
+interface ReingestContractParams {
+  contractAddress: string;
 }
